# swagger.yaml
swagger: "2.0"
info:
  description: "GitHub Release Notification API that allows users to subscribe to email notifications about new releases of a chosen GitHub repository."
  version: "1.0.0"
  title: "GitHub Release Notification API"
host: "releases-api.app"
tags:
  - name: "orchestrator"
    description: "Public front door (:8090) — subscribe + confirm/unsubscribe pages"
  - name: "subscription"
    description: "Subscription service (:8080) — list API (/api prefix)"
schemes:
  - "http"
  - "https"
paths:
  /subscribe:
    post:
      tags:
        - "orchestrator"
      summary: "Subscribe to release notifications"
      description: "Orchestrator (:8090). Runs the subscribe saga: validate + register the repo on GitHub, create the subscription, queue the confirmation email."
      operationId: "subscribe"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Subscription request"
          required: true
          schema:
            $ref: "#/definitions/SubscribeRequest"
      responses:
        "200":
          description: "Subscription successful. Confirmation email sent."
        "400":
          description: "Invalid input (e.g., invalid repo format)"
        "404":
          description: "Repository not found on GitHub"
        "409":
          description: "Email already subscribed to this repository"
  /confirm/{token}:
    get:
      tags:
        - "orchestrator"
      summary: "Confirm email subscription"
      description: "Orchestrator (:8090) HTML result page. Confirms the subscription via NATS request-reply (subscription.confirm). Idempotent: it sets confirmed=true and does NOT consume the token (email scanners prefetch links, so a GET often arrives before the user); re-hitting it is a no-op. The token is reaped by FK cascade when the subscription is deleted on unsubscribe."
      operationId: "confirmSubscription"
      parameters:
        - name: "token"
          in: "path"
          description: "Confirmation token"
          required: true
          type: "string"
      produces:
        - "text/html"
      responses:
        "200":
          description: "Subscription confirmed (result page rendered)"
        "404":
          description: "Token not found"
  /unsubscribe/{token}:
    get:
      tags:
        - "orchestrator"
      summary: "Unsubscribe from release notifications"
      description: "Orchestrator (:8090) HTML result page. Unsubscribes via NATS request-reply (subscription.unsubscribe); the subscription service hard-deletes the row and publishes events.subscription.removed."
      operationId: "unsubscribe"
      parameters:
        - name: "token"
          in: "path"
          description: "Unsubscribe token"
          required: true
          type: "string"
      produces:
        - "text/html"
      responses:
        "200":
          description: "Unsubscribed (result page rendered)"
        "404":
          description: "Token not found"
    post:
      tags:
        - "orchestrator"
      summary: "One-click unsubscribe (RFC 8058)"
      description: "Orchestrator (:8090) List-Unsubscribe-Post one-click endpoint. Same NATS path as the GET; returns status only, no page."
      operationId: "unsubscribeOneClick"
      parameters:
        - name: "token"
          in: "path"
          description: "Unsubscribe token"
          required: true
          type: "string"
      responses:
        "200":
          description: "Unsubscribed successfully"
        "404":
          description: "Token not found"
  /api/subscriptions:
    get:
      tags:
        - "subscription"
      summary: "Get subscriptions for an email"
      description: "Subscription service (:8080). Returns all active subscriptions for the given email address. Gated by X-API-Key when API_KEY is set."
      operationId: "getSubscriptions"
      parameters:
        - name: "email"
          in: "query"
          description: "Email address to look up subscriptions for"
          required: true
          type: "string"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Successful operation - list of subscriptions returned"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Subscription"
        "400":
          description: "Invalid email"
        "401":
          description: "Missing or invalid X-API-Key (when API_KEY is set)"
definitions:
  SubscribeRequest:
    type: "object"
    required:
      - "email"
      - "repo"
    properties:
      email:
        type: "string"
        description: "Email address to subscribe"
      repo:
        type: "string"
        description: "GitHub repository in owner/repo format (e.g., golang/go)"
  Subscription:
    type: "object"
    required:
      - "email"
      - "repo"
    properties:
      email:
        type: "string"
        description: "Email address"
      repo:
        type: "string"
        description: "GitHub repository in owner/repo format"
      confirmed:
        type: "boolean"
        description: "Whether the subscription is confirmed"
      last_seen_tag:
        type: "string"
        description: "Last seen release tag for this repository"